manually revert last 3 commits
