+++ src/datasets/domain/repositories/IDatasetsRepository.ts
@@ -7,7 +7,6 @@ import { DatasetDTO } from '../dtos/DatasetDTO'
 import { DatasetDeaccessionDTO } from '../dtos/DatasetDeaccessionDTO'
 import { MetadataBlock } from '../../../metadataBlocks'
 import { DatasetVersionDiff } from '../models/DatasetVersionDiff'
-import { DatasetVersionInfo } from '../models/DatasetVersionInfo'
 
 export interface IDatasetsRepository {
   getDataset(
@@ -52,5 +51,4 @@ export interface IDatasetsRepository {
     datasetVersionId: string,
     deaccessionDTO: DatasetDeaccessionDTO
   ): Promise<void>
-  getDatasetVersions(datasetId: number | string): Promise<DatasetVersionInfo[]>
 }
