+++ src/datasets/index.ts
@@ -17,7 +17,6 @@ import { PublishDataset } from './domain/useCases/PublishDataset'
 import { UpdateDataset } from './domain/useCases/UpdateDataset'
 import { GetDatasetVersionDiff } from './domain/useCases/GetDatasetVersionDiff'
 import { DeaccessionDataset } from './domain/useCases/DeaccessionDataset'
-import { GetDatasetVersions } from './domain/useCases/GetDatasetVersions'
 
 const datasetsRepository = new DatasetsRepository()
 
@@ -49,7 +48,6 @@ const updateDataset = new UpdateDataset(
   datasetResourceValidator
 )
 const deaccessionDataset = new DeaccessionDataset(datasetsRepository)
-const getDatasetVersions = new GetDatasetVersions(datasetsRepository)
 
 export {
   getDataset,
@@ -64,8 +62,7 @@ export {
   publishDataset,
   createDataset,
   updateDataset,
-  deaccessionDataset,
-  getDatasetVersions
+  deaccessionDataset
 }
 export { DatasetNotNumberedVersion } from './domain/models/DatasetNotNumberedVersion'
 export { DatasetUserPermissions } from './domain/models/DatasetUserPermissions'
