+++ test/environment/.env
@@ -1,6 +1,6 @@
 POSTGRES_VERSION=13
 DATAVERSE_DB_USER=dataverse
 SOLR_VERSION=9.3.0
-DATAVERSE_IMAGE_REGISTRY=ghcr.io
-DATAVERSE_IMAGE_TAG=11188-version-difference-summary
+DATAVERSE_IMAGE_REGISTRY=docker.io
+DATAVERSE_IMAGE_TAG=unstable
 DATAVERSE_BOOTSTRAP_TIMEOUT=5m
